requirements.txt
# Core Django
django==5.0.10
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter
djangorestframework-simplejwt

# Cloudinary
cloudinary
django-cloudinary-storage

# Database
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Environment & Config
python-decouple

# Date & Time
python-dateutil
pytz

# Wagtail CMS
wagtail
wagtail-localize
wagtailmenus

# Images
Pillow
django-imagekit

# Background Tasks
celery
redis
django-celery-beat

# Email
django-anymail

# Calendar Integration
ics
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# Stripe Payment (prepared)
stripe

# API Documentation
drf-spectacular

# Production Server
gunicorn
whitenoise

# Development Tools
ruff
django-debug-toolbar
django-extensions
ipython
